Started work on adding more compiler directives (some not implemented like #ifdef...)
Added initial code for supporting compiler directive for allowing quick index tables.  This required an initializer for the code generator to make sure the mainSymbolTable was passed into everything that required it (generating lookup tables for multiplication needs to add symbols to it).

Added code for allowing compiler directives to know what context they're executing in.

Cleaned up a couple of common string functions.